+++ packages/app/src/api/core-domain/infra/db/raw.ts
@@ -20,3 +20,19 @@ export interface RepresentationEquilibreeRaw {
   siren: string;
   year: number;
 }
+
+export interface OwnershipRequestRaw {
+  asker_email: string;
+  created_at: string;
+  email: string;
+  error_detail?: string;
+  id: string;
+  modified_at: string;
+  siren: string;
+  status: string;
+}
+
+export interface OwnershipRaw {
+  email: string;
+  siren: string;
+}