+++ packages/app/src/common/core-domain/dtos/OwnershipRequestDTO.ts
@@ -1,0 +1,9 @@
+export type OwnershipRequestDTO = {
+  askerEmail: string;
+  createdAt: string;
+  email: string;
+  id: string;
+  modifiedAt: string;
+  siren: string;
+  status: string;
+};